Add ability to supply custom values files to the 'make render' action (#1192)

+++ examples/render-examples.sh
@@ -1,12 +1,37 @@
-#!/usr/bin/env bash
-# Render all the examples in parallel
+#!/bin/bash
+# Purpose: Renders Kubernetes manifests from Helm charts for various examples in parallel.
+# Notes:
+#   - Renders all examples using Helm template.
+#   - Uses the default example values file located in each example directory and optionally additional values files passed as parameters.
+#   - Supports parallel rendering of examples for efficiency.
+#
+# Parameters:
+#   $@: Additional values files to be applied on top of the default example values file for each Helm chart rendering (optional).
+#
+# Usage Examples:
+#   ./render-examples.sh
+#   ./render-examples.sh extra-values.yaml
+#   ./render-examples.sh values1.yaml values2.yaml
 
-SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
+SCRIPT_DIR=$(cd -- "$(dirname -- "${BASH_SOURCE[0]}")" &> /dev/null && pwd)
 
-render_task(){
+render_task() {
   example_dir=$1
   rendered_manifests_dir="${example_dir}rendered_manifests"
-	values_yaml=$example_dir`ls "${example_dir}" | grep values.`
+  default_values_yaml=$(ls "${example_dir}" | grep -m 1 'values.yaml')
+
+  if [ -z "$default_values_yaml" ]; then
+    echo "No default values.yaml found in ${example_dir}"
+    exit 1
+  fi
+
+  # Initialize helm values with the default values file
+  helm_values=("--values" "${example_dir}${default_values_yaml}")
+
+  # Append additional values files if provided
+  for value_file in "${values_files[@]}"; do
+    helm_values+=("--values" "${value_file}")
+  done
 
   # Clear out all rendered manifests
   rm -rf "${rendered_manifests_dir}"
@@ -14,36 +39,40 @@ render_task(){
   # Generate rendered files
   out=$(helm template \
     --namespace default \
-    --values "${values_yaml}" \
+    "${helm_values[@]}" \
     --output-dir "${rendered_manifests_dir}" \
     default helm-charts/splunk-otel-collector)
   if [ $? -ne 0 ]; then
-      echo "$values_yaml FAIL - helm template: $out"
-      exit 1
-  fi
-  # Move the chart renders
-	cp -rp "${rendered_manifests_dir}/splunk-otel-collector/templates/"* $rendered_manifests_dir;
-  if [ $? -ne 0 ]; then
-      echo "${values_yaml} FAIL - Move the chart renders"
-      exit 1
-  fi
-  # Move any subchart renders
-  if [ -d "${rendered_manifests_dir}/splunk-otel-collector/charts/" ]
-  then
-    subcharts_dir="${example_dir}rendered_manifests/splunk-otel-collector/charts"; \
-    subcharts_di=$(find "${subcharts_dir}" -type d -maxdepth 1 -mindepth 1 -exec basename \{\} \;); \
-    for subchart in ${subcharts_di}; do \
-      mkdir -p "${example_dir}rendered_manifests/${subchart}"; \
-      mv "${subcharts_dir}/${subchart}/templates/"* "${example_dir}rendered_manifests/${subchart}"; \
-      if [ $? -ne 0 ]; then
-          echo "${values_yaml} FAIL - Move subchart renders"
-          exit 1
-      fi
-    done; \
+    echo "$default_values_yaml FAIL - helm template: $out"
+    exit 1
   fi
 
-  echo "${values_yaml} SUCCESS"
+  # Move the chart renders
+  cp -rp "${rendered_manifests_dir}/splunk-otel-collector/templates/"* "$rendered_manifests_dir"
+  if [ $? -ne 0 ]; then
+    echo "${default_values_yaml} FAIL - Move the chart renders"
+    exit 1
+  fi
+
+  # Move any subchart renders
+  if [ -d "${rendered_manifests_dir}/splunk-otel-collector/charts/" ]; then
+    subcharts_dir="${rendered_manifests_dir}/splunk-otel-collector/charts"
+    subcharts_di=$(find "${subcharts_dir}" -type d -maxdepth 1 -mindepth 1 -exec basename \{\} \;)
+    for subchart in ${subcharts_di}; do
+      mkdir -p "${rendered_manifests_dir}/${subchart}"
+      mv "${subcharts_dir}/${subchart}/templates/"* "${rendered_manifests_dir}/${subchart}"
+      if [ $? -ne 0 ]; then
+        echo "${default_values_yaml} FAIL - Move subchart renders"
+        exit 1
+      fi
+    done
+  fi
+
+  echo "${default_values_yaml} SUCCESS"
 }
+
+# Collect additional values files passed as arguments
+values_files=("$@")
 
 for example_dir in $SCRIPT_DIR/*/; do
   render_task "${example_dir}" &
@@ -51,7 +80,7 @@ done
 wait # Let all the render tasks finish
 
 for example_dir in $SCRIPT_DIR/*/; do
-  rendered_manifests_dir=$example_dir"rendered_manifests"
+  rendered_manifests_dir="${example_dir}rendered_manifests"
   if [ ! -d "${rendered_manifests_dir}" ]; then
     echo "Examples were rendered, failure occurred"
     exit 1
@@ -63,7 +92,7 @@ for example_dir in $SCRIPT_DIR/*/; do
         rm -rf "${rendered_manifests_dir}/splunk-otel-collector"
     fi
     if [ $? -ne 0 ]; then
-        echo "${values_yaml} FAIL - Temporary space cleanup"
+        echo "${default_values_yaml} FAIL - Temporary space cleanup"
         exit 1
     fi
   fi
